Organize fontconfig wrappers
Each Fc type is split into a separate file. This organization will help
as features are added to the bindings.
